Refactor routes to use layout elements
